adj 3

+++ CppScripts/BBBhw/GPIO.cpp
@@ -705,7 +705,7 @@ unsigned long long int duration_InterruptTag=static_cast<unsigned long long int>
 //cout << "static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(duration_InterruptTag).count()): " << static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(duration_InterruptTag).count()) << endl;
 //this->TimeTaggsLast=static_cast<unsigned long long int>(static_cast<long long int>(this->TimeTaggsInit)+static_cast<long long int>(static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(this->TimePointClockTagPRUinitial-this->TimePointClockTagPRUinitialOld).count())/static_cast<double>(PRUclockStepPeriodNanoseconds)));
 //this->TimePointClockTagPRUinitialOld=this->TimePointClockTagPRUinitial;// Update
-this->TimeTaggsLast=this->TimeTaggsInit+static_cast<unsigned long long int>(static_cast<double>(std::chrono::duration_cast<std::chrono::nanoseconds>(this->TimePointClockTagPRUfinal-this->TimePointClockTagPRUinitialOld).count()-duration_InterruptTag)/static_cast<double>(PRUclockStepPeriodNanoseconds));
+this->TimeTaggsLast=this->TimeTaggsInit+static_cast<unsigned long long int>(static_cast<long double>(std::chrono::duration_cast<std::chrono::nanoseconds>(this->TimePointClockTagPRUfinal-this->TimePointClockTagPRUinitialOld).count()-duration_InterruptTag)/static_cast<long double>(PRUclockStepPeriodNanoseconds));
 this->TimePointClockTagPRUinitialOld=this->TimePointClockTagPRUfinal-std::chrono::nanoseconds(duration_InterruptTag);// Update
 this->TimeTaggsInit=this->TimeTaggsLast;// Update
 //else{Use the latest used, so do not update
